+++ trixnity-utils/src/jvmMain/kotlin/net/folivo/trixnity/utils/ByteBufferExtensions.kt
@@ -1,0 +1,8 @@
+package net.folivo.trixnity.utils
+
+import kotlinx.coroutines.flow.flowOf
+import java.nio.ByteBuffer
+
+fun ByteBuffer.toByteArrayFlow(): ByteArrayFlow {
+    return flowOf(array().copyOf())
+}